+++ build.ps1
@@ -1,59 +1,85 @@
 # thanks to http://ramblingcookiemonster.github.io/Building-A-PowerShell-Module/
 
 Function Resolve-Module {
-    [Cmdletbinding()]
-    param(
-        [Parameter(Mandatory=$true)][String]$Name,
-        [Parameter()][Version]$Version
+    [CmdletBinding()]
+    Param (
+        [Parameter(Mandatory=$true, ValueFromPipeline=$true, ValueFromPipelineByPropertyName=$true)]
+        [String]
+        $Name,
+
+        [Parameter(ValueFromPipelineByPropertyName=$true)]
+        [Version]
+        $Version
     )
 
-    Write-Verbose -Message "Resolving module $Name"
-    $module = Get-Module -Name $Name -ListAvailable
+    Begin {
+        $input_modules = @{}
+    }
 
-    if ($module) {
-        if ($null -eq $Version) {
-            Write-Verbose -Message "Module $Name is present, checking if version is the latest available"
-            $Version = (Find-Module -Name $Name -Repository PSGallery | `
-                Measure-Object -Property Version -Maximum).Maximum
-            $installed_version = ($module | Measure-Object -Property Version -Maximum).Maximum
+    Process {
+        $input_modules.Add($Name, $Version)
+    }
 
-            $install = $installed_version -lt $Version
-        } else {
-            Write-Verbose -Message "Module $Name is present, checking if version matched $Version"
-            $version_installed = $module | Where-Object { $_.Version -eq $Version }
-            $install = $null -eq $version_installed
-        }
+    End {
+        $modules = [System.String[]]$input_modules.Keys
+        $versions = Find-Module -Name $modules -Repository PSGallery
 
-        if ($install) {
-            Write-Verbose -Message "Installing module $Name at version $Version"
-            Install-Module -Name $Name -Force -SkipPublisherCheck -RequiredVersion $Version
+        foreach ($module_name in $modules) {
+            $module_version = $input_modules.$module_name
+            $module = Get-Module -Name $module_name -ListAvailable
+            Write-Verbose -Message "Resolving module $module_name"
+
+            if ($module) {
+                if ($null -eq $module_version) {
+                    Write-Verbose -Message "Module $module_name is present, checking if version is the latest available"
+                    $module_version = ($versions | Where-Object { $_.Name -eq $module_name } | `
+                        Measure-Object -Property Version -Maximum).Maximum
+                    $installed_version = ($module | Measure-Object -Property Version -Maximum).Maximum
+
+                    $install = $installed_version -lt $module_version
+                } else {
+                    Write-Verbose -Message "Module $module_name is present, checking if version matched $module_version"
+                    $version_installed = $module | Where-Object { $_.Version -eq $module_version }
+                    $install = $null -eq $version_installed
+                }
+
+                if ($install) {
+                    Write-Verbose -Message "Installing module $module_name at version $module_version"
+                    Install-Module -Name $module_name -Force -SkipPublisherCheck -RequiredVersion $module_version
+                }
+                Import-Module -Name $module_name -RequiredVersion $module_version
+            } else {
+                Write-Verbose -Message "Module $module_name is not installed, installing"
+                $splat_args = @{}
+                if ($null -ne $module_version) {
+                    $splat_args.RequiredVersion = $module_version
+                }
+                Install-Module -Name $module_name -Force -SkipPublisherCheck @splat_args
+                Import-Module -Name $module_name -Force
+            }
         }
-        Import-Module -Name $Name -RequiredVersion $Version
-    } else {
-        Write-Verbose -Message "Module $Name is not installed, installing"
-        $splat_args = @{}
-        if ($null -ne $Version) {
-            $splat_args.RequiredVersion = $Version
-        }
-        Install-Module -Name $Name -Force -SkipPublisherCheck @splat_args
-        Import-Module -Name $Name -Force
     }
 }
 
+Write-Output -InputObject "Setting up build dependencies"
 Get-PackageProvider -Name NuGet -ForceBootstrap > $null
-
 if ((Get-PSRepository -Name PSGallery).InstallationPolicy -ne "Trusted") {
     Write-Verbose -Message "Setting PSGallery as a trusted repository"
     Set-PSRepository -Name PSGallery -InstallationPolicy Trusted
 }
 
-Resolve-Module -Name Psake
-Resolve-Module -Name PSDeploy
-Resolve-Module -Name Pester
-Resolve-Module -Name BuildHelpers
-Resolve-Module -Name PsScriptAnalyzer
+@(
+    'Psake',
+    'PSCodeCovIo',
+    'PSDeploy',
+    'Pester',
+    'BuildHelpers',
+    'PSScriptAnalyzer'
+) | Resolve-Module
 
-Set-BuildEnvironment
+Write-Output -InputObject "Setting build environment variables"
+Set-BuildEnvironment -ErrorAction SilentlyContinue
 
+Write-Output -InputObject "Starting build"
 Invoke-psake .\psake.ps1
 exit ( [int]( -not $psake.build_success ) )