+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for PInvokeHelper
 
+## v0.2.0 - TBD
+
+* Add pipeline input for `Import-PInvokeMethod` with dynamic type generation
+* Add tests for PowerShell Core
+* Moved across to new CI style with code coverage collection
+* Auto generated documentation with platyPS
+
 ## v0.1.0 - 2019-05-03
 
 * Initial version of the `PInvokeHelper` module